Deterministic plugin ordering
Future features (i.e. CredentialComposer work) rely on deterministic
plugin execution order. The order that plugins are defined is a nice implied
ordering. Unfortunately, decoding the plugins node into a map of maps
hides the order that they were defined.

This change updates the catalog to instead decode configuration from an
HCL ast.Node so that ordering can be preserved.

Signed-off-by: Andrew Harding <azdagron@gmail.com>