Support dedicated non-root user in release images
Natively support a way for users of the release images to run the images
as a non-root user.

SPIRE depends on certain directories to be readable from / writeable to / creatable.
With scratch-based images, the root of the filesystem is empty by default and
only the root user has permission to create directories at the root
level. This means that any directories SPIRE depends on should be
pre-created at build time for some dedicated non-root user.

Create all such directories in the SPIRE Server and Agent image builds.
The directories chosen were ones defined in the
[Linux Filesystem Hierarchy Standard](https://refspecs.linuxfoundation.org/FHS_3.0/fhs/index.html).

Additionally, provide sample configurations that use the paths pre-created
in the images, and add documentation on how to run the images as a non-root user.

Signed-off-by: Ryan Turner <turner@uber.com>