Prevent nil panic from derefeence nil context in authorization middleware (#3829)
* Prevent nil panic from derefeence a nil context in authorization middleware

Nil panics happens authorization middleware's Preprocess function when a nil context is passed down from a failed database operation while fetching entries in function WithCallerEntries. This change addresses this risk by dereference the context and fetch the logger earlier before the ctx variable is mutated.

Extended explanation:

WithCallerEntries returns error when a database operation failure happens. The error itself is not fatal, however, it returns a nil context. Returns zero value when errors is correct behavior. However, the reconcileResult function and opaAuth function bubble up the nil context as if it is populated. In the authorization middleware's Preprocess function, it attempts to retrieve a logger from the context thus cased a nil panic.

Fix:

The fix is simple: since the dereference happens for the purpose of fetching a logger. The logger is prefetched before the context is mutated. This is simplest fix of all the forms. Other alternative fix include introduce a separate variable to hold mutated context and I believe it overcomplicates the code.

Other changes:

This change also include a fix in the authorization_opa.go to explicitly returns nil context in when functions errored out. It doesn't change the logic because it is already returning nil.

Signed-off-by: Troy Dai <git@troydai.com>