Accomodate more plugin config shapes
Recent changes to the plugin config parsing code required to enable
deterministic plugin ordering were too restrictive on the shapes of
plugin configuration, causing some previously-allowed configurations to
become invalid.

This PR changes parsing code to be more flexible when determining order.
It walks the AST and builds up a stack of ast nodes until it has enough
object key's in the stack to determine the plugin type and name.

The actual config parsing is then done separately using the more or less
the same config shape as before, changing to a list of maps of maps to
accomodate the way HCL parses into the AST.

Signed-off-by: Andrew Harding <azdagron@gmail.com>