Introduce CI PR job timeouts
In the past we have observed that GitHub Actions can get stuck on
certain PR jobs, e.g. doing network I/O. The default timeout is 12
hours, which is excessive for the jobs we currently run in the PR status
checks.

Use a reasonably high timeout value that comfortably handles our job
execution times with 50% buffer to account for the jobs taking longer
over time.

Signed-off-by: Ryan Turner <turner@uber.com>