Move release process documentation to new file
The MAINTAINERS.md file has become quite large over time, mostly due to
refinements in the release process. Understanding the details of the
release process are not particularly relevant to understanding the core
duties of a project maintainer, which is the purpose of the
MAINTAINERS.md file. Similarly, maintainers who need to execute the
release process are not concerned with the general documentation about
what it means to be a maintainer on the project.

For these reasons, move the existing release process documentation as is
to a new file in the repository, and reference the new documentation in
MAINTAINERS.md for traceability.

Signed-off-by: Ryan Turner <turner@uber.com>