Fix Linux release artifact checksum files
The checksum files published for Linux release artifacts are published
in a format that is not consumable by tools like
`shasum --check`/`sha256sum --check` because these tools expect two
spaces between the checksum and the filename. Add the missing space to
fix this.

Tested on macOS with `shasum --check <file>` and on Ubuntu with
`sha256sum --check <file>`.

Signed-off-by: Ryan Turner <turner@uber.com>