Directly write the output of the hashing command into the checksum files.
This avoids the cut and pasting that we do to take the hash from the hashing
algorithm and format it with the file from the trimmed full-file path.  It does
so by opening a sub-shell, changing directory to the parent of the file to be
hashed, generating the hash on the file name only (relative to the current
working directory), and then writing the output to the checksum file.

This avoids the need to cut the hashing output to get the hash so we can then
reconstruct the hash line with a file name lacking parent directories.  This
will greatly improve the ability of the --check command to work, as now we don't
alter the format of the sha256sum line.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>