datastore: use UNION ALL for postgres list entries query (#4111)
This cuts the cost in time, as determined by EXPLAIN ANALYZE, in
half. For a database with ~20000 entries and ~50000 selectors the
time goes from ~200ms to ~100ms.

The difference between UNION and UNION ALL is that the former also
does a DISTINCT step on the entries, removing any duplicates. It's
possible I'm missing something simple, but I don't think that actually
does anything for us since most of the sub-queries also return an
unique primary key so they are already unique.

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>
Signed-off-by: Dmitry Gorochovsky <d.goro@yahoo.com>