Systemd workloadattestor closes DBus connection too early
This PR fixes is a bug in the systemd workload attestor. In
`getSystemdUnitInfo`, the DBus connection is closed via a `defer`. But
that connection is required later to retrieve the `ID()` and
`FragmentPath()` in the `Attest` function. This causes attestation
failures with an error message `dbus: connection closed by user`.

The fix extracts the required selector attributes early and avoids
using the dangling connection to DBus.

Signed-off-by: Raghu Kaippully <rkaippully@gmail.com>