Fix reading of unused config keys in HCL parsing
The upgrade to github.com/hashicorp/vault/sdk v0.9.1 introduces a
downstream dependency on a branch of github.com/hashicorp/hcl that has a
different behavior for reading in unrecognized config fields from HCL.
The mainline version of HCL v1 reads these unrecognized fields as a
`[]string` into a struct field tagged with the name `unusedKeys`,
but the new version reads them as a `map[string][]token.Pos` into a struct
field tagged with the name `unusedKeyPositions`.

Fix all the affected spots in the code that handle unrecognized HCL
config fields, and sort the field names for consistent unit test
execution results, since maps with a string key type in Go don't guarantee
any stable ordering of the keys in the map.
Signed-off-by: Ryan Turner <turner@uber.com>