Fix usage of k8s controller-runtime fakeClient
sigs.k8s.io/controller-runtime introduced a breaking change in the
fake client to by default only allow updating of status subresources
in objects shipped in K8s that are known to have the status subresource.
Custom resources that have a status subresource need to register this
object in the fake client builder in order to dynamically update the
status subresource in the test.

This behavior change broke the expectations of the certmanager
upstreamauthority plugin tests that relied on this old behavior of
the fake controller client allowing any resource to have its status
subresource updated. Update the initialization of the fake client to
register the custom `CertificateRequest` object that has a status
subresource, such that it can continue to be updated by tests.
Signed-off-by: Ryan Turner <turner@uber.com>