Fix flaky AttestAgent tests
There is some asynchrony to how gRPC handles Postprocess interceptor
callbacks for bi-directional streaming APIs that results in race
conditions around when audit log messages are emitted by the audit log
middleware. This causes a lot of flakiness of ordering of log messages
received and the sequencing of when those audit logs are received by the
test logrus hook when the tests are run on CI. Resetting the entries
received by the logger hook in-between calls to AttestAgent in the test
is not enough to work around this asynchrony.

Rather than depending on the logs to come in a specific order and
timeframe, instead match all the expected log messages from the gRPC
server in any order rather than just the ones from the last call to
AttestAgent.

Signed-off-by: Ryan Turner <turner@uber.com>