Run upgrade integration test on arm64
The upgrade test couldn't run on arm64 until v1.7.0 because it relies on
the release Docker images from the previous minor release series to all
be available for arm64. The project started publishing Docker release
images for arm64 starting in v1.6.0, so this test can be re-enabled for
arm64 architecture now.

Signed-off-by: Ryan Turner <turner@uber.com>