Use server-generated timestamps in fakedatastore (#4209)
* Use server-generated timestamps in fakedatastore

The server sqlstore was rounding CreatedAt timestamps in the model
object returned back to the client based on the timestamp generated
by the datastore. This rounding logic was added to handle precision
limitations in MySQL 5.7 such that tests could compare created
objects with ones queried from the datastore. Some CI runs of tests
using datastore functionality have uncovered that this rounding logic
does not always produce the same timestamp as the datastore.

Since this rounding is only needed by tests, instead rely on
server-generated timestamps in the test process when using the
fakedatastore. This will guarantee that the CreatedAt timestamps
typically generated by the datastore are always consistent
between the Create and Get/List datastore calls.

Provide this through an undocumented config field in sqlstore that
only is intended to be used by the fakedatastore in tests.

Signed-off-by: Ryan Turner <turner@uber.com>