Fix test on server bind address (#4260)
The test verifies that errors thrown by `net.ResolveTCPAddr()` are
handled and fail server startup. The test case used a string that could
be interpreted as a valid hostname, and `net.ResolveTCPAddr()` allows
hostnames to be used in the address. In some cases, it was noticed that
this test can fail because `net.ResolveTCPAddr()` treats the address as
valid.

Change the address to be something that's not a valid IP address or a
valid hostname to prevent flakiness.

Signed-off-by: Ryan Turner <turner@uber.com>