Don't waste time looking for the entry id when we already know it (#4211)
It looks like in most cases when the agent is looking for a SPIFFE ID it
has to go from selectors to entry to SPIFFE ID and then again from SPIFFE
ID to entry to get the entry id to put in the request (which also involves
fetch all the cached entries, sorting them and then iterating through them
to find one that matches the SPIFFE ID). This seems a bit inefficient.

I think the previous logic was also a bit weird since we lost some precision
going from entry (determined by selectors and contains other information) to
SPIFFE ID and again to picking the first entry that has that SPIFFE ID (which
could be a different one, with different attributes such as `isAdmin`). It
could maybe be considered a security issue, but I don't think anyone would do
something like that (same SPIFFE-ID on admin and non-admin entries).

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>