Clamp bundle refresh interval to a maximum
The refresh hint for a trust domain is useful to know when we expect
to see updates to the bundle on a standard schedule, but there are
situations where we get updates more often, for example when a spire-server
doesn't have any available key (e.g. the first time it's started) and
needs to create a new one and start using it immediately.

Because of this it's useful to clamp the refresh hint to a maxium so we
can learn about those keys sooner rather than later.

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>