Ignore patch updates for GCP SDK modules
GCP SDK modules update quite frequently, but we don't usually make many
changes to the GCP plugins that require SDK updates. We want to stay
relatively up-to-date, but it is very rare that we need a specific
feature/fix from a patch update. Updating to each new patch version is
time consuming from a maintenance standpoint, with little to no benefit.

We already do this for AWS SDK modules, which have a similar cadence of
updates.
Signed-off-by: Ryan Turner <turner@uber.com>