workloadattestor systemd: dbus use of closed network connection
Fixes #4315

The connection to system bus is shared and should not be closed after
use in order to avoid errors on concurrent usage.

It is typical to share the connection in the same process as per this
explanation:
https://github.com/godbus/dbus/issues/179#issuecomment-1006847775

In case of errors, this shared connection will detect that and attempt
to reconnect: https://github.com/godbus/dbus/blob/v5.1.0/conn.go#L124.

Signed-off-by: Raghu Kaippully <rkaippully@gmail.com>