Fix kubectl version detection in integration tests
Newer versions of kubectl have removed the "--short" flag and produce
the output that was once provided by "--short".

To maintain compatabiliy for old versions of kubectl, this change
updates the the common script that detects whether or not kubectl needs
to be downloaded to instead output the version to JSON and uses jq to
extract the version.

Added some helpful diag output as well.

Signed-off-by: Andrew Harding <azdagron@gmail.com>