server: Include CA Serial Number in Subject field
This helps avoid running into a crypto/x509 limit that limits how many
signature checks are allowed to happen to validate a certificate, most
likely to avoid DoS attacks.

Building up a valid certificate chain is done based on the Subject
field, even in the presence of an AKID. If all certificates from the
bundle have the same Subject name, the verifier will attempt to build
up a chain using all of them. Once we have more than 100 certificates
in the bundle, we hit that crypto/x509 limit and start failing to
verify X509-SVIDs.

Additionally this also improves mTLS channel establishment since we now
no longer have to try to build a chain against all certificates from the
bundle, we can easily identify the CA that signed an X509-SVID due to it
having an unique Subject.

Example resulting Subject:
Subject: C = US, O = SPIFFE, serialNumber = 249778918434805435318471934298663515772

An experimental flag is also available to disable this feature in
case there are issues encountered.

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>