Fix: only one of trust_bundle_path, trust_bundle_url, or insecure_bootstrap can be set (#4532)
@mnp reported in issue 4530 that it was possible to set trust_bundle_url
and insecure_bootstrap in the Agent configuration. There was a test for
this case. However, the test was just checking if there was an error.
There was an error but not the expected one. This commit also adds
expectErrorContains to the test case struct so tests can check the
expected error message. Also, more tests added.

Signed-off-by: Matteus Silva <silvamatteus@lsd.ufcg.edu.br>
Signed-off-by: Faisal Memon <fymemon@yahoo.com>