Fix SyncAuthorizedEntries test race
The SyncAuthorizedEntries test is set up to optionally send a request,
depending on the expectations of the test case. The "no caller ID" and
"fetcher fails" test cases are currently set up to send the request on
the stream. However, in this instances, the handler exits before reading
the request off the stream due to the respective expected failures.
Timing conditions still allowed the requests to be sent successfully
most of the time but caused intermittent failures, particularly when
being run in resource constrained environments (e.g. GH actions).

This change fixes the test cases by not attempting to send the request
on the stream.

Signed-off-by: Andrew Harding <azdagron@gmail.com>