Fix racy manager NotifiesBundleLoaded test
This test is racy because it currently relies on timing of several
goroutines and context cancellation.

The notifier doing context cancellation doesn't seem to test anything
useful. I don't recall why it was even added.

Getting rid of the context cancellation and simply asserting that the
function returns nil when the notifier returns nil seems good enough.

Signed-off-by: Andrew Harding <azdagron@gmail.com>