Log SPIFFE ID for X.509-SVIDs signed in BatchNewX509SVID (#4902)
The audit log emitted on calls to BatchNewX509SVID doesn't include the
SPIFFE ID of the X.509-SVIDs that are signed during the API handler
execution. It's valuable to include the SPIFFE ID in this log message
for traceability and auditing purposes. The SPIFFE ID in signed
X.509-SVIDs is currently only included in a DEBUG level log in the
server CA.

Signed-off-by: Ryan Turner <turner@uber.com>