Try out github.com/christophebedard/dco-check
The DCO app currently used in the repository doesn't support GitHub
merge queues. Try out this other DCO check script which will run as our
own managed job that we can configure to run for PRs added to the merge
queue.

We can try running this check in parallel to the existing check to
ensure it's stable before trying to rely on it.

Signed-off-by: Ryan Turner <turner@uber.com>