lru cache: Only send notifications to the new subscriber
When there's a new subscription to the cache we shouldn't notify all existing
subscriptions. That causes them to receive updates even if there was no
change to the SVIDs or bundles.

This can cause issues with users that trigger side effects, such as
restarting a service, when they receive updates on the X509 stream.

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>