No longer test against MySQL 5.7
MySQL 5.7 is EOL as of October 2023.

Stop using MySQL 5.7 in integration tests, since it is not easy for the
project to support end-of-life database versions. Also log a warning
when MySQL 5.7 is used with SPIRE.

Importantly, this commit does not remove any existing support for MySQL
5.7. SPIRE may continue to work for some time with MySQL 5.7, but the
project makes no compatibility guarantees with MySQL 5.7 anymore, and
any existing compatibility with MySQL 5.7 may break at any time without
notice.

It is recommended that all SPIRE users using MySQL as a SQL backend
upgrade to MySQL 8.0.

Signed-off-by: Ryan Turner <turner@uber.com>