Use timestamp invalidate kubelet response cache
Changes the K8s workload attestor to invalidate the cache response based
on a timestamp instead of kicking off a goroutine.

While this means that the plugin will always have a cached response in
memory, it simplifies the code for both the plugin and tests. In
practice, only an idle agent would waste memory holding onto the
response.

Signed-off-by: Andrew Harding <azdagron@gmail.com>