agent: Accept registration entries with RevisionNumber 0 (#5680)
When using the SyncAuthorizedEntries API we verify that the
RevisionNumber of the entries is at least 1. Unfortunately,
the RevisionNumber as assigned by spire-server starts at 0,
leading to the following error:
```
WARN[0004] Received malformed entry revision from SPIRE server; are the server and agent versions compatible?  entry_id=f1a26ecf-1043-4905-b5d9-789e60f88695 revision_number=0 subsystem_name=manager
````

I'm guessing the check is there in the unlikely case of a wrap-around,
so I'm keeping it and verifying that the RevisionNumber is positive.

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>