Use a temporary file for the fakedatastore (#5815)
With a in memory sqlite datastore the database gets cleaned up when all connections to it get closed. If that happens, the next connection to be opened will see an empty database and usually error out.

I've noticed this in tests ocassionally and this seems to make running the db using  not fail.

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>