Extract the trust bundle code from agent/cli/run
The trust bundle fetching code is only available from the cli run file.
It will need to be in a more common location for reattestation code to
come. We move the existing code to a common location with as few changes
as possible. This will make future reviews easier as changes will not
involve moving large blocks of code at the same time as changing
behavior for reattestation support.

Signed-off-by: Kevin Fox <Kevin.Fox@pnnl.gov>