Delete some dead code
I ran
    go run golang.org/x/tools/cmd/deadcode@latest -test ./...
and manually reviewed the result.

I selected a subset of "dead code" (never-called functions) to delete. I
did not delete some other functions that seemed like they could still be
useful in future.

One function flagged as dead was "DeltaSecretsTest" in
pkg/agent/endpoints/sdsv3/handler_test.go, which seems like it was a
misspelled function that was intended as a test case. I tried renaming
it to "TestDeltaSecrets" but it turned out to be broken and hard to fix.
Since it only tests that an RPC method is unimplemented, I opted to just
remove the function instead.

Signed-off-by: Carlo Teubner <cteubner1@bloomberg.net>