TestAttestAgent: fix flakeyness
This flakeyness was reported previously in #5774 with improvements made
in #5815, but was still happening.

Hopefully fix it, by not assuming that the gRPC Send() message must not
return an error; it is legitimately allowed to return io.EOF, in the
case where the server has returned an application-level error, which
we can then obtain via a Recv() call.

See https://pkg.go.dev/google.golang.org/grpc#ClientStream.SendMsg

Signed-off-by: Carlo Teubner <cteubner1@bloomberg.net>