golangci-lint: upgrade to v2 & fix flakey test (#6064)
* golangci-lint: upgrade to v2

- Upgrade golangci-lint to the latest version, v2.1.6.
- Migrate to new config file format as required with v2.
- Run golangci-lint via "go run" not "go tool"; the docs explicitly
  recommend against the latter, and it was needlessly polluting our
  go.mod file.
- With the new way, the golangci-lint version is maintained in
  .spire-tool-versions, consistent with other similar tools.
- While we're at it, enable a few more linters/checks.
- Address all new linter complaints.
- Simplify some error return code patterns
- golangci-lint: enable more revive rules

* TestAttestAgent: fix flakeyness

This flakeyness was reported previously in #5774 with improvements made
in #5815, but was still happening.

Hopefully fix it, by not assuming that the gRPC Send() message must not
return an error; it is legitimately allowed to return io.EOF, in the
case where the server has returned an application-level error, which
we can then obtain via a Recv() call.

See https://pkg.go.dev/google.golang.org/grpc#ClientStream.SendMsg

Signed-off-by: Carlo Teubner <cteubner1@bloomberg.net>