ci: fix some test failures due to clocks (#6201)
* Wait for ticker to wait before continuing with the test

Otherwise we might advance the clock before the ticker starts and we block until the test timeouts.

Signed-off-by: Sorin Dumitru <sorin@returnze.ro>

* Unflake TestRunUpdateCacheTaskDoesFullUpdate test

I think what happens here is that we don't get any guarantee about the order the tickers tick when we advance the clock the second time.

Signed-off-by: Sorin Dumitru <sorin@returnze.ro>

* Use Eventually for checks against background goroutine

The goroutine might take longer than the test to run, in which case the assertions will fail.

Signed-off-by: Sorin Dumitru <sorin@returnze.ro>

---------

Signed-off-by: Sorin Dumitru <sorin@returnze.ro>