Fix unit tests that improperly use assert.Equal()
A couple tests are using `assert.Equal()` inside a
`require.Eventually()`. `assert.Equal()` will cause the test to fail if
the assertion is not met. This conflicts with the intention of retrying
the equality check in the test.

These tests started failing due to a new behavior in
`github.com/stretchr/testify` v1.11.0 that started executing the
function passed to `require.Eventually()`/`assert.Eventually()` without
any initial delay: https://github.com/stretchr/testify/pull/1427

The tests were previously passing by chance due to the timing of how
long `github.com/stretchr/testify` was waiting to try the function in
`Eventually()` for the first time.

Signed-off-by: Ryan Turner <ryan.turner253@icloud.com>