add joinTokenFile option for agent configuration
This is so it's easier to integrate with modern credential provider
solutions like systemd-creds,  kubernetes secrets  or the various
container-storage-interface based credential providers.

All these provide secrets by mounting them into a location on disk.
This is because env vars are not suitable for secrets as they can leak
across fork() boundaries and are generally not treated as private
by the OS and libraries.

Signed-off-by: Arian van Putten <arian.vanputten@gmail.com>