OIDC discovery provider parm handling (#6297)
* feat: add error handling for unexpected positional arguments in OIDC discovery provider

Implemented a check in the main function to handle unexpected positional arguments, providing a clear error message and usage instructions. This enhances the robustness of the command-line interface for the OIDC discovery provider.

Signed-off-by: Jenova Marie <jenova@recoverysky.org>

* refactor: remove commented-out code for unexpected positional arguments in OIDC discovery provider

Removed the commented-out check for unexpected positional arguments in the main function to clean up the code. This change streamlines the implementation while maintaining the existing error handling functionality.

Signed-off-by: Jenova Marie <jenova@recoverysky.org>

* fix: update usage message format in OIDC discovery provider

Modified the usage message in the OIDC discovery provider to use "Usage of" instead of "Usage:". This change ensures consistency in the output and improves clarity in error handling for unexpected arguments.

Signed-off-by: Jenova Marie <jenova@recoverysky.org>

* Add gosec exceptions for safe test usage of os.Args[0]

Signed-off-by: Jenova Marie <jenova@recoverysky.org>

---------

Signed-off-by: Jenova Marie <jenova@recoverysky.org>
Co-authored-by: Sorin Dumitru <sorin@returnze.ro>