Add a variant of the X.509 CA metric with absolute expiration time, not TTL (#6303)
* Add support for float64 gauge metrics

Signed-off-by: Kim Vandry <vandry@TZoNE.ORG>

* Add a variant of the X.509 CA metric with absolute expiration time, not TTL.

In Prometheus at least, a metric with the TTL is not usable because it is
not known when the metric was set and thus what time the TTL is relative
to (it is possible to convey the timestamp when the metric was set using
an optional extended scrape format, but Spire does not export in that
format and even if it did the resulting metric would more difficult to
work with in Prometheus). It is better to publish the absolute expiration
time and let the monitoring system subtract the current time to get a
TTL if it wants.

I left the existing TTL-based metric in place rather than deleting it in
case that way is useful or preferred by one of the other supported
telemetry exporters, but if it isn't then I would recommend deprecating it.

Signed-off-by: Kim Vandry <vandry@TZoNE.ORG>

---------

Signed-off-by: Kim Vandry <vandry@TZoNE.ORG>
Co-authored-by: Sorin Dumitru <sorin@returnze.ro>