fix: fetch federated trust bundles concurrently
SPIRE agent periodically syncs entries, federated bundles and SVIDs from
the SPIRE server. This happens every 5 seconds with a backoff. There is
a 30 second timeout for this process. If there are many federations for
the entries allowed for the agent, this timeout can be exceeded.

This change adds a concurrent fetch for federated bundles to reduce the
chance of hitting the timeout.

Fixes: #6490