fix: fetch federated trust bundles concurrently
SPIRE agent periodically syncs entries, federated bundles and SVIDs from
the SPIRE server. This happens every 5 seconds with a backoff. There is
a 30 second timeout for this process. If there are many federations for
the entries allowed for the agent, this timeout can be exceeded.

This change adds a concurrent fetch for federated bundles to reduce the
chance of hitting the timeout. It also adds a unit test to cover the
case when there are more federated bundles than the number of workers.

Fixes: #6490
Signed-off-by: Mark Goddard <mark@cofide.io>