Use returned SPIFFE ID returned with JWT-SVID for caching
We use our cached entry for requesting a new svid which may be different from
the one that the server has. It may happen that entry gets updated around the
same time that a cached SVID is renewed, including changing the SPIFFE ID. In
this case the SPIFFE ID of the returned SVID may be different from the one of
the entry we have.

Because of the above we shouldn't use the SPIFFE ID of our entry for caching
the SVIDs. The NewJWTSVID API also returns the SPIFFE ID of the issued SVID so
we can use that for caching purposes.

X509-SVID doesn't have this issue since there we cache using the entry ID.

Signed-off-by: Sorin Dumitru <sorin@returnze.ro>